+++ react-accordion/my-app/src/Accordion.tsx
@@ -1,26 +1,22 @@
-import { TopicCard } from "./TopicCard";
+import { useState } from 'react';
+import { TopicCard } from './TopicCard';
 
-const topics = [
-  {
-    id: 1,
-    title: 'Hypertext Markup Language',
-    content:
-      'Hypertext Markup Language (HTML) is the standard markup language for documents designed to be displayed in a web browser. It can be assisted by technologies such as Cascading Style Sheets (CSS) and scripting languages such as JavaScript.',
-  },
-  {
-    id: 3,
-    title: 'Cascading Style Sheets',
-    content:
-      'Cascading Style Sheets (CSS) is a style sheet language used for describing the presentation of a document written in a markup language like HTML. CSS is a cornerstone technology of the World Wide Web, alongside HTML and JavaScript.',
-  },
-  {
-    id: 9,
-    title: 'JavaScript',
-    content:
-      'JavaScript (/ˈdʒɑːvəˌskrɪpt/), often abbreviated as JS, is a high-level, interpreted scripting language that conforms to the ECMAScript specification. JavaScript has curly-bracket syntax, dynamic typing, prototype-based object-orientation, and first-class functions.',
-  },
-];
+type Props = {
+  topics: {
+    id: number;
+    title: string;
+    content: string;
+  }[];
+};
 
-export function Accordion() {
-  return <TopicCard />
+export function Accordion({ topics }: Props) {
+  const [topic, setTopic] = useState<number | null>(null);
+
+  function handleClick(id: number) {
+    setTopic((prevId) => (prevId === id ? null : id));
+  }
+
+  return (
+    <TopicCard topics={topics} onClick={handleClick} selectedTopic={topic} />
+  );
 }
